+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 ### History
 
+#### 04/09/25
+- continue working at delete old rows flow, need to test
+- new method at google service added to get the index of a row only by date
+- TODO:
+ - continue checking if ai fails 
+ - add all tests for google module
+
 #### 03/09/25
 - delete old rows flow defined 
 - started working at delete old rows flow
 - test cases for use-cases added
 - date.service tests refactored
-- TODO:
- - continue checking if ai fails 
- - add all tests for google module
 
 
 #### 31/08/25